+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/runtime/ServiceRuntimeAdapter.java
@@ -174,7 +174,6 @@ public class ServiceRuntimeAdapter {
    * @throws TransactionExecutionException if the transaction execution failed
    * @see ServiceRuntime#executeTransaction(int, String, int, byte[], Fork, int, HashCode,
    *      PublicKey)
-   *      Review: patch
    * @see com.exonum.binding.core.transaction.Transaction
    */
   void executeTransaction(int serviceId, String interfaceName, int txId, byte[] arguments,